Cleanup according to analyzers
After applying clang-tidy and codemodel of Qt Creator, apply a couple of
changes to make the code more clean.

Remaining items are kept because either they touch released code or no
mean to optimize (i.e. padding on QSharedData based classes).

+++ src/mqtt/qmqttcontrolpacket.cpp
@@ -125,7 +125,7 @@ QByteArray QMqttControlPacket::serializePayload() const
 {
     QByteArray data;
     // Add length
-    quint32 msgSize = m_payload.size();
+    quint32 msgSize = quint32(m_payload.size());
     if (msgSize > 268435455)
         qWarning("Publishing a message bigger than maximum size!");
     do {
